+++ src/PimEnterprise/Bundle/VersioningBundle/Denormalizer/Flat/ProductValue/AttributeOptionDenormalizer.php
@@ -13,6 +13,8 @@ namespace PimEnterprise\Bundle\VersioningBundle\Denormalizer\Flat\ProductValue;
 
 use Pim\Bundle\CatalogBundle\Entity\Repository\AttributeOptionRepository;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Attribute option flat denormalizer used for following attribute types:
@@ -41,7 +43,16 @@ class AttributeOptionDenormalizer extends AbstractValueDenormalizer
      */
     public function denormalize($data, $class, $format = null, array $context = array())
     {
+        if ($data === null || $data === '') {
+            return null;
+        }
+
+        $resolver = new OptionsResolver();
+        $this->configContext($resolver);
+        $context = $resolver->resolve($context);
+
         $value = $context['value'];
+
         $option = $this->findEntity(
             $this->prepareOptionCode($data, $value)
         );
@@ -69,10 +80,19 @@ class AttributeOptionDenormalizer extends AbstractValueDenormalizer
      *
      * @param string $identifier
      *
-     * @return AttributeOption
+     * @return \Pim\Bundle\CatalogBundle\Entity\AttributeOption
      */
     protected function findEntity($identifier)
     {
         return $this->repository->findByReference($identifier);
+    }
+
+    /**
+     * Define context requirements
+     * @param OptionsResolverInterface $resolver
+     */
+    protected function configContext(OptionsResolverInterface $resolver)
+    {
+        $resolver->setRequired(['value']);
     }
 }