+++ src/PimEnterprise/Bundle/WorkflowBundle/Presenter/ProductValue/FilePresenter.php
@@ -11,7 +11,7 @@
 
 namespace PimEnterprise\Bundle\WorkflowBundle\Presenter\ProductValue;
 
-use Pim\Bundle\CatalogBundle\Model\AbstractProductMedia;
+use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use PimEnterprise\Bundle\WorkflowBundle\Presenter\TwigAware;
 use PimEnterprise\Bundle\WorkflowBundle\Presenter\TwigAwareInterface;
@@ -34,7 +34,7 @@ class FilePresenter implements ProductValuePresenterInterface, TwigAwareInterfac
     public function supports(ProductValueInterface $value)
     {
         return 'pim_catalog_file' === $value->getAttribute()->getAttributeType()
-            && $value->getData() instanceof AbstractProductMedia;
+            && $value->getData() instanceof ProductMediaInterface;
     }
 
     /**